Corrected the middleware order
The middleware for running a delete query was run before the middleware for an update query, I switched the order.  The original docs caused the following problem: When I run a delete query, but the record is already 'soft-deleted', the query (which has been converted to an update query in the middleware) will still run and return that record(s) are deleted because the update-middleware could not run to filter the query on deleted = false because it was running before the delete middleware.